Add model serialization tests

+++ Sources/FinnhubSwift/Models/EconomicCalendar.swift
@@ -6,7 +6,7 @@ public struct EconomicEvent: Codable, Equatable {
     var estimate: Double?
     var event: String
     var impact: String
-    var prev: Int
+    var prev: Double
     var time: String
     var unit: String
 
@@ -28,20 +28,20 @@ public struct EconomicCalendar: Codable, Equatable {
     public static func == (lhs: EconomicCalendar, rhs: EconomicCalendar) -> Bool {
         return lhs.economicCalendar == rhs.economicCalendar
     }
-    
+
     /*
-     
-     {
-        "economicCalendar":[{
-            "actual":1133,
-            "country":"NZ",
-             "estimate":null,
-             "event":"New Zealand-Tourist arrivals-Estimated Migrant Arrivals",
-             "impact":"low",
-             "prev":-172,
-             "time":"2020-10-11 21:45:00",
-             "unit":"Number of"
-        }]
-     }
-    */
+
+      {
+         "economicCalendar":[{
+             "actual":1133,
+             "country":"NZ",
+              "estimate":null,
+              "event":"New Zealand-Tourist arrivals-Estimated Migrant Arrivals",
+              "impact":"low",
+              "prev":-172,
+              "time":"2020-10-11 21:45:00",
+              "unit":"Number of"
+         }]
+      }
+     */
 }
